enable TCP_NODELAY on tcp connections
Nagle's algorithm is designed to collect groups of writes from
applications that many make write() calls with a few bytes at a time
into a single packet.

dctx is not that dumb, so introducing those delays is always a
hinderance.  The number of 3-node roundtrips that the python code can do
without Nagle's algorithm is near 24000.  With Nagle's algorithm, that
drops to as low as 100.